+++ src/main/kotlin/me/evgen/advbot/db/local/LocalStorage.kt
@@ -2,6 +2,8 @@ package me.evgen.advbot.db.local
 
 import chat.tamtam.botsdk.model.ChatId
 import chat.tamtam.botsdk.model.UserId
+import chat.tamtam.botsdk.model.prepared.Chat
+import me.evgen.advbot.model.AdPlatform
 import me.evgen.advbot.model.entity.TempAdvert
 import java.util.concurrent.atomic.AtomicLong
 
@@ -9,9 +11,7 @@ class LocalStorage {
     private val idGenerator = AtomicLong(0L)
     private val tempAdvertsByUserId = mutableMapOf<Long, TempAdvert>()
 
-    //TODO: сделать мапу по Chat вместо ChatId
-    private val chatNamesMap: MutableMap<ChatId, String> = mutableMapOf()
-    private val userChatsMap: MutableMap<UserId , MutableSet<ChatId>> = mutableMapOf()
+    private val chatsMap: MutableMap<UserId, MutableSet<AdPlatform>> = mutableMapOf()
 
     fun getNextId(): Long = idGenerator.getAndIncrement()
 
@@ -31,34 +31,49 @@ class LocalStorage {
         tempAdvertsByUserId.remove(userId)
     }
 
-    fun addChatName(id: ChatId, name: String) {
-        chatNamesMap[id] = name
-    }
+    fun addChat(userId: UserId, chat: Chat) {
+        if (chatsMap.containsKey(userId)) {
+            chatsMap[userId]!!.add(createAdPlatformFromChat(chat))
 
-    fun getChatName(id: ChatId): String {
-        return if (chatNamesMap.containsKey(id)) {
-            chatNamesMap[id]!!
         } else {
-            "Not found"
-        }
-    }
-
-    fun addChat(userId: UserId, id: ChatId) {
-        if (userChatsMap.containsKey(userId)) {
-            userChatsMap[userId]!!.add(id)
-        } else {
-            val chatSet = mutableSetOf<ChatId>().apply {
-                add(id)
+            val chatSet = mutableSetOf<AdPlatform>().apply {
+                add(createAdPlatformFromChat(chat))
             }
-            userChatsMap[userId] = chatSet
+            chatsMap[userId] = chatSet
         }
     }
 
     fun removeChat(user: UserId, id: ChatId) {
-        userChatsMap[user]?.removeIf { it.id == id.id }
+        //TODO
     }
 
-    fun getChats(user: UserId): Set<ChatId> {
-        return userChatsMap[user] ?: setOf()
+    fun getChats(userId: UserId): Set<Chat> {
+        val chats = mutableSetOf<Chat>()
+        if (chatsMap.containsKey(userId)) {
+            for (entry in chatsMap[userId]!!) {
+                chats.add(entry.chat)
+            }
+        }
+        return chats
+    }
+
+    fun getPlatform(userId: UserId, chatId: Long): AdPlatform? {
+        return chatsMap[userId]?.find { it.chat.chatId.id == chatId }
+    }
+
+    fun tagSwitchPlatform(userId: UserId, adPlatform: AdPlatform, tag: String) {
+        chatsMap[userId]?.first { it.getChatId() == adPlatform.getChatId() }?.apply {
+            if (!tags.contains(tag)) {
+                tags.add(tag)
+            } else tags.remove(tag)
+        }
+    }
+
+    private fun createAdPlatformFromChat(chat: Chat): AdPlatform {
+        return AdPlatform(
+            chat,
+            mutableSetOf(),
+            false
+        )
     }
 }
