+++ src/main/kotlin/me/evgen/advbot/model/state/PlatformAccessSwitchState.kt
@@ -3,13 +3,12 @@ package me.evgen.advbot.model.state
 import chat.tamtam.botsdk.client.RequestsManager
 import chat.tamtam.botsdk.state.CallbackState
 import me.evgen.advbot.BotController
-import me.evgen.advbot.getUser
+import me.evgen.advbot.db.DBSessionFactoryUtil
 import me.evgen.advbot.getUserId
-import me.evgen.advbot.storage.LocalStorage
 
 class PlatformAccessSwitchState(timestamp: Long, private val chatId: Long) : BaseState(timestamp), CustomCallbackState {
     override suspend fun handle(callbackState: CallbackState, prevState: BaseState, requestsManager: RequestsManager) {
-        val adPlatform = LocalStorage.getPlatform(callbackState.getUserId(), chatId)
+        val adPlatform = DBSessionFactoryUtil.localStorage.getPlatform(callbackState.getUserId(), chatId)
         if (adPlatform == null) {
             "Ошибка! Нет такой платформы.".answerNotification(callbackState.getUserId(), callbackState.callback.callbackId, requestsManager)
             return
@@ -21,7 +20,7 @@ class PlatformAccessSwitchState(timestamp: Long, private val chatId: Long) : Bas
             callbackState.getUserId(), callbackState.callback.callbackId, requestsManager)
 
         val newState = PlatformAccessSettingsState(timestamp, chatId)
-        BotController.moveTo(newState, callbackState.getUser()) {
+        BotController.moveTo(newState, callbackState.getUserId().id) {
             newState.handle(callbackState, prevState, requestsManager)
         }
     }
